+++ ethereum/eip712/encoding_legacy.go
@@ -11,7 +11,7 @@ import (
 	txTypes "github.com/cosmos/cosmos-sdk/types/tx"
 	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 	apitypes "github.com/ethereum/go-ethereum/signer/core/apitypes"
-	evmos "github.com/evmos/evmos/v19/types"
+	evmos "github.com/evmos/os/types"
 )
 
 type aminoMessage struct {