+++ testutil/constants.go
@@ -1,0 +1,12 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
+
+package testutil
+
+const (
+	// ExampleChainID provides a chain ID that can be used in tests
+	ExampleChainID = "evmos_9000-1"
+
+	// ExampleDenom provides an example denom for use in tests
+	ExampleAttoDenom = "aevmos"
+)