+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog was created using the `clu` binary
 
 ### Improvements
 
+- (all) [#15](https://github.com/evmos/os/pull/15) Add general types and utils.
 - (proto) [#14](https://github.com/evmos/os/pull/14) Add Protobufs and adjust scripts.
 - (eip-712) [#13](https://github.com/evmos/os/pull/13) Add EIP-712 package.
 - (ci) [#12](https://github.com/evmos/os/pull/12) Add CI workflows, configurations, Makefile, License, etc.